Nonparametric EDMA fit

Estimate parameters of a structured SigmaK, partly based on Hu (2007).

## read pattern file
read_pattern(file, ...)

## estimation
SigmaK_fit(object, pattern, check_pattern = TRUE, ...)

## methods
SigmaK(object, ...)
sensitivity(object, ...)
# S3 method for edma_fit_p
print(x, truncate = 40, ...)
# S3 method for edma_fit_p
SigmaK(object, ...)
# S3 method for edma_fit_p
sensitivity(object, m=10, ...)

Arguments

object, x
EDMA fit object.

file
the name of the file which the data are to be read from using read.csv or read_excel.

truncate
numeric, number of characters to print for the object title.

pattern
a pattern matrix, see Details.

check_pattern
logical, if pattern should be checked for identifiability.

m
positive integer, number of times SigmaK estimation is to be replicated to assess sensitivity.

...
other arguments passed to methods.

Details

The pattern matrix describes the unknowns in the SigmaK matrix. Diagonal must be filled. Upper and lower diagonal values must be symmetric or NA. See Examples.

Value

read_pattern reads in a csv or xls/xlsx file with the pattern matrix. See Examples.

SigmaK_fit estimates the SigmaK matrix and attaches that to the EDMA fit object. The estimated SigmaK matrix can be accessed by the SigmaK method. Estimation is done by optim.

sensitivity is used to re-estimate SigmaK matrix multiple times with random starting values for numerical optimization. When estimates vary while the loss function value is the same, it is usually a sign of non-identifiability. The return object has columns starting with par_* that are the parameters according to pattern matrix, value is the loss function value evaluated at the estimates.

References

Hu, L., 2007. Euclidean Distance Matrix Analysis of Landmarks Data: Estimation of Variance. Thesis, Master of Science in Statistics, Department of Mathematical and Statistical Sciences, University of Alberta, Edmonton, Alberta, Canada. Pp. 49.

See also

Examples

## read pattern matrix in cs or xls/xlsx format
read_pattern(
    system.file("extdata/example.csv", package="EDMAinR"))

#>   a   b   c  
#> a "1" "2" NA 
#> b "2" "1" NA 
#> c NA  NA  "1"

read_pattern(
    system.file("extdata/example.xlsx", package="EDMAinR"))

#>   a   b   c  
#> a "1" "2" NA 
#> b "2" "1" NA 
#> c NA  NA  "1"

## here is a pattern matrix
m <- matrix(c(
    "a", NA, NA, NA,
    NA, "a", NA, NA,
    NA,  NA, "b", NA,
    NA,  NA, NA, "b"
), 4, 4, byrow=TRUE)
parm <- c(a=0.25, b=0.35)
m

#>      [,1] [,2] [,3] [,4]
#> [1,] "a"  NA   NA   NA  
#> [2,] NA   "a"  NA   NA  
#> [3,] NA   NA   "b"  NA  
#> [4,] NA   NA   NA   "b" 

## simulate data, so we know what the true values were
M <- structure(c(-2.5, 7.5, -2.5, -2.5, -7.5, 2.5, 2.5, 4.5),
    .Dim = c(4L, 2L))
SigmaK <- EDMAinR:::.vec2mat(parm, EDMAinR:::.mat2fac(m))
sim <- edma_simulate_data(n=500, M, SigmaK)
dimnames(M) <- dimnames(sim$data[[1L]])
rownames(SigmaK) <- rownames(m) <- rownames(sim$data[[1L]])
colnames(SigmaK) <- colnames(m) <- rownames(sim$data[[1L]])

## nonparametric fit
fit <- edma_fit(sim)
## parametric fit
o <- SigmaK_fit(fit, m)
o

#> EDMA parametric fit: Simulated landmark data
#> Call: SigmaK_fit(object = fit, pattern = m)
#> 2 dimensions, 4 landmarks, 500 replicates, no bootstrap

## estimates
cbind(true=parm, est=o$results$par)

#>   true       est
#> a 0.25 0.2511162
#> b 0.35 0.3341726

SigmaK(o)

#>           L1        L2        L3        L4
#> L1 0.2511162 0.0000000 0.0000000 0.0000000
#> L2 0.0000000 0.2511162 0.0000000 0.0000000
#> L3 0.0000000 0.0000000 0.3341726 0.0000000
#> L4 0.0000000 0.0000000 0.0000000 0.3341726

## sensitivity
s <- sensitivity(o)
summary(s)

#>      par_a            par_b            value      
#>  Min.   :0.2491   Min.   :0.3248   Min.   :1.201  
#>  1st Qu.:0.2520   1st Qu.:0.3313   1st Qu.:1.201  
#>  Median :0.2536   Median :0.3317   Median :1.201  
#>  Mean   :0.2534   Mean   :0.3319   Mean   :1.201  
#>  3rd Qu.:0.2540   3rd Qu.:0.3332   3rd Qu.:1.201  
#>  Max.   :0.2605   Max.   :0.3362   Max.   :1.201